change ReadBits from std::integral to base::bits::UnsignedInteger
Bug: None
Cq-Include-Trybots: luci.chromium.try:linux-cast-x64-rel
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6857623
Commit-Queue: Weidong Liu <weidongliu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1506246}

+++ media/base/container_names.cc
@@ -476,7 +476,7 @@ static bool CheckH261(const uint8_t* buffer, int buffer_size) {
 
     // Skip any extra insertion information. Since this is open-ended, if we run
     // out of bits assume that the buffer is correctly formatted.
-    int extra = ReadBits(&reader, 1);
+    uint8_t extra = ReadBits(&reader, 1);
     while (extra == 1) {
       if (!reader.SkipBits(8))
         return seen_start_code;
@@ -487,7 +487,7 @@ static bool CheckH261(const uint8_t* buffer, int buffer_size) {
     // Next should be a Group of Blocks start code. Again, if we run out of
     // bits, then assume that the buffer up to here is correct, and the buffer
     // just happened to end in the middle of a header.
-    int next;
+    uint16_t next;
     if (!reader.ReadBits(16, &next))
       return seen_start_code;
     RCHECK(next == 1);
